Fixed a bug in board hash values after pawn promotion
Also, this commit and the previous commit have changed the way the promotion piece is represented
The promotion piece is now always listed as the white version of that piece, and is given the right color when the promotion occurs.
This change was made for compatibility with polyglot format opening books.